Add optional script to replicate a local repo
Restic is known to make a large number of API requests to services
like B2 where these requests are chargeable. The simplest workaround
is to perform the backup to a local repository and then use rclone
to sync the repository to the remote location. `push.sh` implements
this step of the process.