Use `--group-by host` for transient files
In my use-case for transient files, the files included in the
snapshot change every day. A typical snapshot might be

```
.../log/archive/2020-10,oct/production.log.12,mon
.../log/production.log
.../log/production.log.last
```

restic's default `--group-by` will put each snapshot in a separate
group, and so all of them will be retained. Using `--group-by host`
will group them all together, and the `--tag transient` ensures
we are only targeting the transient snapshots if a single repository
is being used.